requirements.txt
streamlit
langchain
langchain-community
langchain-groq
sentence-transformers
torch
faiss-cpu
pymupdf
pydantic
matplotlib
python-dotenv
pillow
